Move service/dynamodb.go to aws/dynamodb
